Move ReactGridStore interface to separate file, update imports.
