+++ src/utils/getStyledRangesCSS.tsx
@@ -1,8 +1,8 @@
-import { ReactGridStore } from "./reactGridStore";
 import { StyledRange, StyledRangesCSS } from "../types/PublicModel";
 import { getNumericalRange } from "./getNumericalRange";
 import { InternalStyledRange } from "../types/InternalModel";
 import { createStyledRangesCSS } from "./createStyledRangesCSS";
+import { ReactGridStore } from "../types/ReactGridStore.ts";
 
 
 export function getStyledRangesCSS(store: ReactGridStore, styledRanges: StyledRange[]): StyledRangesCSS {
