+++ src/utils/reactGridStore.ts
@@ -1,64 +1,9 @@
-import { StoreApi, create, createStore, useStore } from "zustand";
+import { create, createStore, StoreApi, useStore } from "zustand";
 import { CellSelectionBehavior } from "../behaviors/CellSelectionBehavior";
 import { DefaultBehavior } from "../behaviors/DefaultBehavior";
-import { NumericalRange } from "../types/CellMatrix";
-import { FocusedCell, IndexedLocation, PaneName } from "../types/InternalModel";
-import { Cell, CellMap, Column, Range, Row, SpanMember, StyledRange } from "../types/PublicModel";
+import { Range, StyledRange } from "../types/PublicModel";
 import { isSpanMember } from "./cellUtils";
-import { RowMeasurement } from "../types/RowMeasurement";
-import { ColumnMeasurement } from "../types/ColumnMeasurement";
-import { Behavior, BehaviorId } from "../types/Behavior";
-
-export interface ReactGridStore {
-  rows: Row[];
-  readonly setRows: (rows: Row[]) => void;
-  readonly getRowAmount: () => number;
-  columns: Column[];
-  readonly setColumns: (columns: Column[]) => void;
-  readonly getColumnAmount: () => number;
-  cells: CellMap;
-  readonly setCells: (cellMap: CellMap) => void;
-  readonly getCellByIds: (
-    rowId: ReactGridStore["rows"][number]["id"],
-    colId: ReactGridStore["rows"][number]["id"]
-  ) => Cell | null;
-  readonly getCellByIndexes: (rowIndex: number, colIndex: number) => Cell | null;
-  readonly getCellOrSpanMemberByIndexes: (rowIndex: number, colIndex: number) => Cell | SpanMember | null;
-
-  rowMeasurements: RowMeasurement[];
-  readonly setRowMeasurements: (rowMeasurements: RowMeasurement[]) => void;
-  colMeasurements: ColumnMeasurement[];
-  readonly setColMeasurements: (colMeasurements: ColumnMeasurement[]) => void;
-
-  paneRanges: Record<PaneName, NumericalRange>;
-  readonly setPaneRanges: (paneRanges: Record<PaneName, NumericalRange>) => void;
-
-  focusedLocation: IndexedLocation;
-  readonly setFocusedLocation: (rowIndex: number, colIndex: number) => void;
-  readonly getFocusedCell: () => FocusedCell | null;
-
-  selectedArea: NumericalRange;
-  readonly setSelectedArea: (selectedArea: NumericalRange) => void;
-
-  currentlyEditedCell: IndexedLocation;
-  readonly setCurrentlyEditedCell: (rowIndex: number, colIndex: number) => void;
-
-  reactGridRef?: HTMLDivElement;
-  readonly assignReactGridRef: (reactGridRef?: HTMLDivElement) => void;
-
-  hiddenFocusTargetRef?: HTMLDivElement;
-  readonly assignHiddenFocusTargetRef: (hiddenFocusTargetRef?: HTMLDivElement) => void;
-
-  styledRanges: StyledRange[];
-  readonly setStyledRanges: (styledRanges: StyledRange[]) => void;
-  readonly getStyledRanges: (range?: Range) => StyledRange[] | [];
-
-  /* == Behaviors == */
-  behaviors: Record<BehaviorId, Behavior>;
-  currentBehavior: Behavior;
-  readonly setBehaviors: (behaviors: Record<BehaviorId, Behavior>) => void;
-  readonly getBehavior: (behaviorId: BehaviorId) => Behavior;
-}
+import { ReactGridStore } from "../types/ReactGridStore.ts";
 
 type ReactGridStores = Record<string, StoreApi<ReactGridStore>>;
